+++ tests/test_njitted.py
@@ -24,7 +24,7 @@ def test_amat_x():
     z = np.arange(1, grid.nCz+1)[::-1]/10
     res_x = np.outer(np.outer(x, y), z).ravel()
     freq = 0.319
-    model = utils.Model(grid, res_x, 0.8*res_x, 2*res_x, freq=freq)
+    model = utils.Model(grid, res_x, 0.8*res_x, 2*res_x)
 
     # Create a source field
     sfield = utils.get_source_field(grid=grid, src=src, freq=freq)
@@ -36,15 +36,15 @@ def test_amat_x():
     rr1 = utils.Field(grid)
     njitted.amat_x(
             rr1.fx, rr1.fy, rr1.fz, efield.fx, efield.fy, efield.fz,
-            model.eta_x, model.eta_y, model.eta_z, model.smu0, model.zeta,
+            model.eta_x, model.eta_y, model.eta_z, sfield.smu0, model.zeta,
             grid.hx, grid.hy, grid.hz)
 
     # amat_x - alternative
     rr2 = utils.Field(grid)
     alternatives.alt_amat_x(
             rr2.fx, rr2.fy, rr2.fz, efield.fx, efield.fy, efield.fz,
-            model.smu0*model.eta_x, model.smu0*model.eta_y,
-            model.smu0*model.eta_z, model.zeta, grid.hx, grid.hy, grid.hz)
+            sfield.smu0*model.eta_x, sfield.smu0*model.eta_y,
+            sfield.smu0*model.eta_z, model.zeta, grid.hx, grid.hy, grid.hz)
 
     # Check all fields (ex, ey, and ez)
     assert_allclose(-rr1, rr2, atol=1e-23)
@@ -102,7 +102,7 @@ def test_gauss_seidel():
         res_y = 0.5*np.arange(grid.nC)+1
         res_z = 2*np.arange(grid.nC)+1
 
-        model = utils.Model(grid, res_x, res_y, res_z, freq)
+        model = utils.Model(grid, res_x, res_y, res_z)
 
         # Initialize source field.
         sfield = utils.get_source_field(grid, src, freq)
@@ -111,7 +111,7 @@ def test_gauss_seidel():
         efield = solver.solver(grid, model, sfield, maxit=2, verb=1)
 
         inp = (sfield.fx, sfield.fy, sfield.fz, model.eta_x, model.eta_y,
-               model.eta_z, model.smu0, model.zeta, grid.hx, grid.hy, grid.hz,
+               model.eta_z, sfield.smu0, model.zeta, grid.hx, grid.hy, grid.hz,
                nu)
 
         # Get result from `gauss_seidel`.