+++ CHANGELOG.rst
@@ -5,10 +5,18 @@ Changelog
 latest - will be *v0.9.0*
 -------------------------
 
+Backwards incompatible in many senses, but mainly for ``emg3d.utils.Model``.
+
 - Utilities:
 
   - ``Fields`` and returned receiver-arrays (``EMArray``) both have amplitude
     (``.amp``) and phase (``.pha``) attributes.
+  - ``Fields`` have attributes containing frequency-information (``freq``,
+    ``smu0``).
+  - New class ``SourceField``; a subclass of ``Field``, adding ``vector`` and
+    ``v{x,y,z}`` attributes for the real valued source vectors.
+  - The ``Model`` is not frequency-dependent any longer and does NOT take
+    a ``freq``-parameter any more.
   - ``data_write`` automatically removes ``_vol`` from ``TensorMesh``
     instances, and ``_eta_{x,y,z}``, ``_zeta`` from ``Model`` instances. This
     makes the archives smaller, and they are not required, as they are simply